account creation is complete
currently I am using a "custom database" aka all of the users are stored in json format in a txt file on the server side. It is really scuffed but I don't expect to need a more professional setup and mongo was giving me issues. Next step is to allow signing in and have a user cookie so I can track what users are doing what. If I want to track actual user data I will need to set up a proper database and find a host, but for now I just plan to store secret Santa requests.